tests(helpers): separate server mock functions (#13630)
https://konghq.atlassian.net/browse/KAG-5254
